Add presentation class and example document
This commit introduces the presentation document class for LaTeX
and an example presetation to show how to use it. For now I only
focused on providing the version 2 (v2) of the official University
layout, the version 1 (v1) would require additional work that will
be done in the future. Right now it supports Polish (pl) and English
(en) background version and both 4:3 and 16:9 aspect ratio.