Improve HR Excellence in Research icon
This commit changes the HR Excellence in Research icon, which is
visible on the title slide. The original embedded image in University
template has poor quality, however there is a LaTeX template, provided
by Wojciech Myszka [1], which is in well scalable format and looks
much better after embedding it in the final document.

Myszka's template provides the same other background files as
I extracted already from the official template, but I decided
to keep the current mechanism for fetching those for visibility
and any knowledge-hungry visitors.

[1] https://kmim.wm.pwr.edu.pl/myszka/